pyproject.toml
[project]
name = "relayops-ai"
version = "0.1.0"
description = "RelayOps AI workflow and integration platform demo"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.116.1",
  "uvicorn==0.35.0",
  "httpx==0.28.1",
  "alembic==1.16.5",
  "sqlalchemy==2.0.43",
  "prometheus-client==0.23.1",
  "opentelemetry-api==1.37.0",
  "opentelemetry-sdk==1.37.0",
  "opentelemetry-exporter-otlp-proto-http==1.37.0",
  "sentry-sdk==2.38.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
